# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface Alert {
  id: string;
  title: string;
  message: string;
  severity: AlertSeverity;
  location_name: string;
  /** @nullable */
  related_report_id?: string | null;
  read: boolean;
  created_at: Date;
}