# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CountPoint } from './countPoint';

export interface Analytics {
  total_reports: number;
  verification_rate: number;
  resolution_rate: number;
  high_risk_areas: number;
  /** @nullable */
  average_resolution_hours: number | null;
  reports_over_time: CountPoint[];
  by_category: CountPoint[];
  by_status: CountPoint[];
  risk_distribution: CountPoint[];
}