# lib/api-zod/src/generated/types/countPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CountPoint {
  label: string;
  count: number;
}