# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRole } from './dashboardSummaryRole';
import type { Report } from './report';

export interface DashboardSummary {
  role: DashboardSummaryRole;
  total_reports: number;
  pending: number;
  verified: number;
  resolved: number;
  high_risk_areas: number;
  recent_reports: Report[];
  demo_mode: boolean;
}