# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  id: string;
  location_name: string;
  prediction_type: string;
  probability: number;
  expected_start: Date;
  expected_end: Date;
  contributing_factors: string[];
  recommendation: string;
}