# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CityPulse urban intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScoreRiskLevel } from './riskScoreRiskLevel';

export interface RiskScore {
  id: string;
  location_name: string;
  latitude: number;
  longitude: number;
  traffic_risk: number;
  flood_risk: number;
  infrastructure_risk: number;
  pollution_risk: number;
  verified_report_risk: number;
  overall_risk: number;
  risk_level: RiskScoreRiskLevel;
  calculated_at: Date;
}